backend/pyproject.toml
[project]
name = "orcatrading-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.4",
  "SQLAlchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "redis>=5.0",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000

